codestyle(RectangleFactory.cs): Убрано лишнее преобразование в int

+++ src/Programming/Programming/Model/Classes/Geometry/RectangleFactory.cs
@@ -14,8 +14,8 @@ static class RectangleFactory
 
         int rectangleWidth = random.Next(10, 100);
         int rectangleHeight = random.Next(10, 100);
-        int centerX = random.Next((int)(15 + (rectangleWidth / 2)), (int)(canvasWidth - 15 - (rectangleWidth / 2)));
-        int centerY = random.Next((int)(15 + (rectangleHeight / 2)), (int)(canvasHeight - 15 - (rectangleHeight / 2)));
+        int centerX = random.Next(15 + (rectangleWidth / 2), canvasWidth - 15 - (rectangleWidth / 2));
+        int centerY = random.Next(15 + (rectangleHeight / 2), canvasHeight - 15 - (rectangleHeight / 2));
 
         Rectangle rectangle = new Rectangle(rectangleWidth, rectangleHeight, new Point2D(centerX, centerY));
